refactor(core): submission 域迁移到 src/domains 并接入门面
