+++ noj-core/src/app.ts
@@ -10,9 +10,9 @@ import adminTrainings from "./domains/catalog/routes/admin-trainings.ts";
 import tags from "./domains/catalog/routes/tags.ts";
 import problems from "./domains/catalog/routes/problems.ts";
 import checkin from "./domains/identity/routes/checkin.ts";
-import queue from "./routes/queue.ts";
-import submissions from "./routes/submissions.ts";
-import selfTests from "./routes/self-tests.ts";
+import queue from "./domains/submission/routes/queue.ts";
+import submissions from "./domains/submission/routes/submissions.ts";
+import selfTests from "./domains/submission/routes/self-tests.ts";
 import users from "./domains/identity/routes/users.ts";
 import rankings from "./routes/rankings.ts";
 import conversations from "./domains/messaging/routes/conversations.ts";
