+++ noj-core/src/domains/gateway/index.ts
@@ -2,6 +2,7 @@ export {
   createUserLlmProvider,
   deleteUserLlmProvider,
   getLlmProviderById,
+  getUserLlmProvider,
   listUserLlmProviders,
   LlmGatewayError,
   testUserLlmProvider,
