+++ noj-core/src/domains/query/index.ts
@@ -1,1 +1,10 @@
-export { getMyRanking } from "../../services/rankings.ts";
+export {
+  applyNewResult,
+  getCachedTodayStats,
+  getCachedTotalStats,
+} from "../../services/stats-cache.ts";
+export {
+  getGlobalRankings,
+  getMyRanking,
+  refreshRankingsView,
+} from "../../services/rankings.ts";
