+++ noj-core/src/domains/submission/index.ts
@@ -1,5 +1,7 @@
+export * from "./services/submissions/submissions.ts";
 export {
-  createArtifactSubmission,
-  createSubmission,
-  listSubmissions,
-} from "../../services/submissions/submissions.ts";
+  DEFAULT_ARTIFACT_MAX_SIZE_BYTES,
+  getArtifactHardLimit,
+} from "./services/submissions/artifact-submissions.ts";
+export * from "./services/queue.ts";
+export * from "./services/self-tests.ts";
