+++ noj-core/src/domains/submission/routes/self-tests.ts
@@ -1,11 +1,11 @@
 import { Hono } from "hono";
-import { authMiddleware } from "../middleware/auth.ts";
-import { parseJsonBody } from "../lib/request.ts";
-import { BadRequestError } from "../lib/errors.ts";
-import { enforceSelfTestRateLimit } from "../lib/hardening-rate-limit.ts";
-import { resolveProblem } from "../lib/problem-resolve.ts";
+import { authMiddleware } from "../../../middleware/auth.ts";
+import { parseJsonBody } from "../../../lib/request.ts";
+import { BadRequestError } from "../../../lib/errors.ts";
+import { enforceSelfTestRateLimit } from "../../../lib/hardening-rate-limit.ts";
+import { resolveProblem } from "../../../lib/problem-resolve.ts";
 import { createSelfTest, getSelfTest } from "../services/self-tests.ts";
-import type { SelfTestInput } from "../types/self-tests.ts";
+import type { SelfTestInput } from "../../../types/self-tests.ts";
 
 type Env = {
   Variables: {
