+++ noj-core/src/routes/admin/index.ts
@@ -14,7 +14,7 @@ import type { AuthEnv } from "../../middleware/auth.ts";
 import { adminMiddleware, authMiddleware } from "../../middleware/auth.ts";
 import adminUsers from "../../domains/identity/routes/admin-users.ts";
 import adminProblems from "../../domains/catalog/routes/admin-problems.ts";
-import adminSubmissions from "./admin-submissions.ts";
+import adminSubmissions from "../../domains/submission/routes/admin-submissions.ts";
 import adminContests from "../../domains/contest/routes/admin-contests.ts";
 import adminJudgeImages from "./admin-judge-images.ts";
 import adminSettings from "./admin-settings.ts";
