+++ noj-core/src/routes/sse.ts
@@ -5,8 +5,8 @@ import type { OptionalAuthEnv } from "../middleware/auth.ts";
 import { Channels, onEvent } from "../lib/event-bus.ts";
 import { createSseStream } from "../lib/sse-stream.ts";
 import { replaySseEvents } from "../lib/sse-events.ts";
-import { getSubmission } from "../services/submissions/submissions.ts";
-import { getQueueOverview } from "../services/queue.ts";
+import { getSubmission } from "../domains/submission/index.ts";
+import { getQueueOverview } from "../domains/submission/index.ts";
 import { checkPermission } from "../lib/permissions.ts";
 import {
   getCachedTodayStats,
